irk3d/requirements.txt
asgiref==3.7.2
django-debug-toolbar==4.1.0
Django==4.2.4
Pillow==9.5.0
pip==23.2.1
setuptools==68.0.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.41.1
django-ckeditor==6.7.0

